// types/questionnaire/personal-info.ts
// Personal information page types
export interface personalInfoFormData {
  firstName: string;
  middleName: string;
  lastName: string;
  dateOfBirth: string;
  languageFluency: string;
  nationality: string;
  domicileStatus: string;
  currentProfession: string;
  street1: string;
  city: string;
  county: string;
  postcode: string;
  marriageDate: string;
}